pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anomaly-detection-api"
version = "0.1.0"
description = "Visual anomaly detection API using PatchCore on MVTec AD, with FastAPI serving, heatmap visualization, and web interface."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Yaniss Amazouz", email = "yaniss91600@gmail.com" }]
keywords = ["anomaly-detection", "patchcore", "computer-vision", "mvtec", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "numpy>=1.26",
    "opencv-python-headless>=4.9",
    "pillow>=10.2",
    "scikit-learn>=1.4",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pre-commit>=3.7",
    "httpx>=0.27",
]
demo = [
    "streamlit>=1.32",
    "plotly>=5.20",
]

[project.urls]
Repository = "https://github.com/YanissAmz/anomaly-detection-api"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
